Move build-related deps to devDependencies
NPM audit lists a bunch of things like DOS vulnerabilities and so on for
libs used in tools that are not receiving network requests (therefore
DOS is irrelevant) because they are used only at build-time.  I've tried
to identify which all these might be apart from just the
create-react-app scripts, and move them to devDependencies so that I
won't get these false alarms any more.