Add nvmrc locking nodejs to the current stable/LTS version
It technically does work with newer versions too (I'm using node v19
locally) but I don't want to get caught off guard or have to test this
too often, if it works with LTS it'll be fine.  The reason this is
needed is to force Netlify to build with a newer version.  Currently the
build is failing due to a syntax error.  It's using node v10 but it
works fine on my machine so I'm guessing that version is too old.  Why
Netlify isn't using the latest stable version despite its own
documentation is a mystery but anyway, the same documentation says you
can override it using .node-version, .nvmrc or an environment variable
so hopefully it'll eat up this non-numeric .nvmrc version specification
and leave me alone forever.